move more support functions out to helpers.py
